# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "学伴 API",
    "description": "学伴（XueBan）AI 学习软件后端服务",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/diagnosis/start": {
      "post": {
        "tags": [
          "diagnosis"
        ],
        "summary": "开启入学水平诊断（F-01）",
        "description": "创建诊断并按初始难度给出第一题。",
        "operationId": "start_v1_diagnosis_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosisStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/diagnosis/{exam_id}/answer": {
      "post": {
        "tags": [
          "diagnosis"
        ],
        "summary": "提交诊断作答（自适应选题）",
        "description": "判定作答、更新画像与难度，并返回下一题或收卷。",
        "operationId": "answer_v1_diagnosis__exam_id__answer_post",
        "parameters": [
          {
            "name": "exam_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Exam Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosisAnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisAnswerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/diagnosis/{exam_id}/report": {
      "get": {
        "tags": [
          "diagnosis"
        ],
        "summary": "诊断报告（含知识点掌握度）",
        "description": "返回诊断报告（未结束时 409）。",
        "operationId": "report_v1_diagnosis__exam_id__report_get",
        "parameters": [
          {
            "name": "exam_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Exam Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/micro-lessons": {
      "post": {
        "tags": [
          "micro-lessons"
        ],
        "summary": "创建知识点微课（异步生成，F-16）",
        "description": "创建微课并投递异步生成任务（不阻塞请求）。",
        "operationId": "create_lesson_v1_micro_lessons_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MicroLessonCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MicroLessonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/micro-lessons/{lesson_id}/audio": {
      "get": {
        "tags": [
          "micro-lessons"
        ],
        "summary": "微课音频地址（可在线播放）",
        "description": "返回预签名音频地址；未就绪返回 409。",
        "operationId": "get_audio_v1_micro_lessons__lesson_id__audio_get",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MicroLessonAudioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiagnosisAnswerRequest": {
        "properties": {
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          },
          "answer": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Answer"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "answer"
        ],
        "title": "DiagnosisAnswerRequest",
        "description": "提交作答。"
      },
      "DiagnosisAnswerResponse": {
        "properties": {
          "exam_id": {
            "type": "string",
            "format": "uuid",
            "title": "Exam Id"
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "correct_answer": {
            "type": "string",
            "title": "Correct Answer"
          },
          "progress": {
            "$ref": "#/components/schemas/DiagnosisProgress"
          },
          "finished": {
            "type": "boolean",
            "title": "Finished"
          },
          "next_question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiagnosisQuestionOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "exam_id",
          "is_correct",
          "correct_answer",
          "progress",
          "finished"
        ],
        "title": "DiagnosisAnswerResponse",
        "description": "作答判定与下一题。"
      },
      "DiagnosisProgress": {
        "properties": {
          "answered": {
            "type": "integer",
            "title": "Answered"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "answered",
          "total"
        ],
        "title": "DiagnosisProgress",
        "description": "进度。"
      },
      "DiagnosisQuestionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "stem": {
            "type": "string",
            "title": "Stem"
          },
          "qtype": {
            "type": "string",
            "title": "Qtype"
          },
          "options": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "difficulty": {
            "type": "integer",
            "title": "Difficulty"
          },
          "knowledge_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Knowledge Points"
          }
        },
        "type": "object",
        "required": [
          "id",
          "stem",
          "qtype",
          "difficulty"
        ],
        "title": "DiagnosisQuestionOut",
        "description": "诊断题目（不下发答案）。"
      },
      "DiagnosisReportResponse": {
        "properties": {
          "exam_id": {
            "type": "string",
            "format": "uuid",
            "title": "Exam Id"
          },
          "finished": {
            "type": "boolean",
            "title": "Finished"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/MasteryPointResponse"
            },
            "type": "array",
            "title": "Points"
          },
          "has_data": {
            "type": "boolean",
            "title": "Has Data"
          }
        },
        "type": "object",
        "required": [
          "exam_id",
          "finished",
          "summary",
          "points",
          "has_data"
        ],
        "title": "DiagnosisReportResponse",
        "description": "诊断报告。"
      },
      "DiagnosisStartRequest": {
        "properties": {
          "subject": {
            "type": "string",
            "maxLength": 32,
            "title": "Subject",
            "default": "math"
          },
          "stage": {
            "type": "string",
            "maxLength": 20,
            "title": "Stage",
            "default": "junior"
          },
          "target_count": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 20.0,
            "title": "Target Count",
            "description": "题量 20~30",
            "default": 25
          }
        },
        "type": "object",
        "title": "DiagnosisStartRequest",
        "description": "开启诊断。"
      },
      "DiagnosisStartResponse": {
        "properties": {
          "exam_id": {
            "type": "string",
            "format": "uuid",
            "title": "Exam Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "$ref": "#/components/schemas/DiagnosisProgress"
          },
          "question": {
            "$ref": "#/components/schemas/DiagnosisQuestionOut"
          }
        },
        "type": "object",
        "required": [
          "exam_id",
          "status",
          "progress",
          "question"
        ],
        "title": "DiagnosisStartResponse",
        "description": "开启诊断响应。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MasteryPointResponse": {
        "properties": {
          "knowledge_point_id": {
            "type": "string",
            "format": "uuid",
            "title": "Knowledge Point Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "mastery": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Mastery"
          },
          "level": {
            "type": "string",
            "title": "Level",
            "description": "red/yellow/green 预警分级"
          },
          "total_attempts": {
            "type": "integer",
            "title": "Total Attempts"
          },
          "correct_attempts": {
            "type": "integer",
            "title": "Correct Attempts"
          },
          "last_practiced_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Practiced At"
          }
        },
        "type": "object",
        "required": [
          "knowledge_point_id",
          "code",
          "name",
          "subject",
          "stage",
          "mastery",
          "level",
          "total_attempts",
          "correct_attempts"
        ],
        "title": "MasteryPointResponse",
        "description": "单个知识点掌握度。"
      },
      "MicroLessonAudioResponse": {
        "properties": {
          "lesson_id": {
            "type": "string",
            "format": "uuid",
            "title": "Lesson Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "mime": {
            "type": "string",
            "title": "Mime"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "lesson_id",
          "url",
          "mime",
          "expires_in"
        ],
        "title": "MicroLessonAudioResponse",
        "description": "微课音频地址。"
      },
      "MicroLessonCreateRequest": {
        "properties": {
          "knowledge_point_id": {
            "type": "string",
            "format": "uuid",
            "title": "Knowledge Point Id"
          }
        },
        "type": "object",
        "required": [
          "knowledge_point_id"
        ],
        "title": "MicroLessonCreateRequest",
        "description": "创建微课。"
      },
      "MicroLessonResponse": {
        "properties": {
          "lesson_id": {
            "type": "string",
            "format": "uuid",
            "title": "Lesson Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "char_count": {
            "type": "integer",
            "title": "Char Count"
          },
          "retries": {
            "type": "integer",
            "title": "Retries"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "lesson_id",
          "title",
          "status",
          "char_count",
          "retries"
        ],
        "title": "MicroLessonResponse",
        "description": "微课状态。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}